+++ changelog.md
@@ -1,3 +1,17 @@
+## 1.2.0 - 2026-06-05
+### Added
+* **`AlignExpr`** — new public symbol
+* **`AlignExprCarry`** — new public symbol
+* **`CarryBackwardExpr`** — new public symbol
+* **`CarryForwardExpr`** — new public symbol
+* **`CarryNearestExpr`** — new public symbol
+### Changed
+* **`Query`** — new optional field(s): align
+### Beta
+* **`LabReportResult`** — field(s) removed: isSensitive
+* **`LabReportResultIsSensitive`** — public symbol removed
+* **`LabReportResultSensitivity`** — new public symbol
+
 ## 1.1.0 - 2026-05-27
 ### Added
 * **`LabTestsClient.updateOrder()`** — new method to update a modifiable order's scheduled activation date via PATCH, accepting an `UpdateOrderBody` with `orderId` and optional `activateBy`.